Move ItemSource, TrafficItem and DeviceItem to new item.rs file.
These types define the items shown in tree views, and the API for
fetching and querying them.

This API is effectively a separate layer on top of the capture
database, so it makes sense to split it out.

Move the Bytes type to util.rs and the InterfaceEpNum type to
usb.rs, as these need to be used by both capture.rs and item.rs.
