+++ src/row_data.rs
@@ -6,7 +6,7 @@ use gtk::subclass::prelude::*;
 #[cfg(any(test, feature="record-ui-test"))]
 use gtk::prelude::Cast;
 
-use crate::capture::{TrafficItem, DeviceItem};
+use crate::item::{TrafficItem, DeviceItem};
 use crate::tree_list_model::ItemNodeRc;
 
 /// Trait implemented by each of our row data types.
@@ -61,7 +61,7 @@ mod imp {
     use gtk::glib::{self, subclass::prelude::*};
     use std::cell::RefCell;
 
-    use crate::capture::{TrafficItem, DeviceItem};
+    use crate::item::{TrafficItem, DeviceItem};
     use crate::tree_list_model::ItemNodeRc;
 
     /// Define the inner type to be used in the GObject type system.
