+++ src/tree_list_model.rs
@@ -15,7 +15,8 @@ use gtk::gio::prelude::ListModelExt;
 use derive_more::AddAssign;
 use itertools::Itertools;
 
-use crate::capture::{CaptureReader, ItemSource};
+use crate::capture::CaptureReader;
+use crate::item::ItemSource;
 use crate::model::GenericModel;
 use crate::row_data::GenericRowData;
 use crate::item_widget::ItemWidget;
